Main Window Basic Design
The Main Window of the application now has a basic design. In my view I
have tackled most of the road blocks and hence will leave the project
for a few days until I work on other parts of bookeeping to get the data
to this UI layer.

Main Window:
1. Menu
2. Left Filesystem Nav.
3. Right Main Filesystem View
4. Toolbar & Status Bar