New Signal Slot Mechanics
Twig will not have one file for declaring and emiting all of its
signals. The class is singleton, hence all of the classes will have the
same instance and will use that instance to connect to the signals.

Note: Signals under directory and file have not been ported for now.
They will be done later.