DefaultView child tree done
The tree of the child is completely done now. The hierarcy is placed
correclty. Note that the records are not brought until the user opens
the directory. This has been made intetional so that at the start of the
program, the use should not wait for long time because a lot of calls
will be made to get the whole filesystem structure.

On other hand, if the user connection is very slow, opening each
directory will take time. In my defense, the JSON for one directory is
much smaller and should not take much time. Also, we can try and cache
the fetched JSON later to fasten the process.