Initial attempt to filesystem change
This is the initial attempt to diff the local and remote xray. The idea
is that when the user will choose to see what changes they have made in
the filesystem, the local xray and the remote xray must be compared and
presented to the user in nice format. Then the user may wish to commit
or rollback those changes.