Gui Correction: File/Directory name total visibility
Previously the name of the entity was not visible because of the size of
the layout. The size have been set to a minumum of 200. Also the text
was being shown at (0,0) from the bottom right. It has been adjusted to
shown (5, -10) so that the text is a little above the bottom to be
clearly visible.